test: move unit tests to auth transformer
