+++ packages/amplify-graphql-index-transformer/package.json
@@ -38,7 +38,6 @@
     "graphql-transformer-common": "4.23.2"
   },
   "devDependencies": {
-    "@aws-amplify/graphql-auth-transformer": "^0.9.2",
     "@aws-cdk/assert": "~1.124.0"
   },
   "jest": {
